# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto API Specification
  description: |-
    This OpenAPI document describes a set of interfaces to allow a WASAPhoto user 
    to interact with the WASAPhoto backend, WASAPhoto is a new social network that
    allows you to share your best moments with friends!
  version: "v1.0.0"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      description: |
        Bearer token for the WASAPhoto API.
      scheme: bearer
      bearerFormat: /^[A-F0-9]{64}$/i
  schemas:
    Username:
      title: Username
      type: object
      description: |-
        The username of the user, it must begin with an alphabetical character 
        and can be composed of a number of alphanumeric characters and underscores, 
        with a minimum length of 3 and a maximum length of 32.
      properties:
        username-string:
          type: string
          description: The username of the user
          pattern: ^[a-zA-Z][a-zA-Z0-9_]{2,32}$ # we don't want to accept any whitespace or weird characters.
          example: "Dario_Loi_123"
          minLength: 2
          maxLength: 32
    SHA256hash:
      title: SHA256hash
      type: object
      description: |-
        A generic SHA256 hash, used throughout the API to identify users, photos and more.
      properties:
        hash:
          type: string
          description: A unique identifier, in this case a SHA256 hash.
          pattern: ^[a-fA-F0-9]{64}$
          example: 985981bedf5cc6956330a1e4c39526a8839e8a3ec3adf0f2e3eedd59c359a58d
          minLength: 64
          maxLength: 64
    Photo:
      title: Photo
      format: binary
      type: string
      minLength: 0
      maxLength: 32000000 # 32MB
      description: |-
        A Base64 encoded photo, used generally all over the place, this is WASA*Photo* after all.
      pattern: ^[A-Za-z0-9+/=]*$
...
    IDList:
      title: IDList
      type: object
      description: |-
        A list of SHA256 hashes, used to represent a list of photos, posts or users.
      properties:
        ids:
          type: array
          description: The list of SHA256 hashes
          minItems: 0
          maxItems: 4294967295
          items:
            $ref: "#/components/schemas/SHA256hash"
...
    Error:
      title: Error
      type: object
      description: |-
        An error message, it contains a code and a message.
      properties:
        code:
          type: integer
          description: The error code
          example: 400
        message:
          type: string
          description: The error message
          example: Bad request
          pattern: ^[^\\]*?$
          minLength: 1
          maxLength: 256
...
paths:
...
  /users/{user_name}/profile/photos:
    parameters:
      - name: user_name
        in: path
        description: The username of the user whose photos are being requested
        required: true
        schema:
          $ref: "#/components/schemas/Username"
    get:
      operationId: getUserPhotos
      summary: Get a user's photos
      tags:
        - "photos"
      description: |-
        Get a user's photos, including their description and the date they were posted,
        useful to populate the feed of their profile.

      responses:
        "200":
          description: |-
            The user's photos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IDList"
        "404":
          description: |-
            The user does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: |-
            Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /users/{user_name}/following/{followed_name}:
    parameters:
      - name: user_name
        in: path
        description: The user's name
        required: true
        schema:
          $ref: "#/components/schemas/Username"
      - name: followed_name
        in: path
        description: The target's username
        required: true
        schema:
          $ref: "#/components/schemas/Username"
    put:
      operationId: followUser
      summary: Follow a user
      description: |-
        Follow a user
      tags:
        - "followers"
      requestBody:
        description: |-
          The user to follow
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Username"
      security:
        - bearerAuth: []
      responses:
        "204":
          description: |-
            The user has been followed
        "400":
          description: |-
            The follow request is ill-formed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: |-
            The user does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: |-
            The user is not correctly authenticated (the given ID does not match the user's ID)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: |-
            The user is already following the target
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: unfollowUser
      summary: Unfollow a user
      description: |-
        Unfollow a user that the user has been following, requires 
        proper authentication from the user.
      tags:
        - "followers"
      security:
        - bearerAuth: []
      responses:
        "204":
          description: |-
            The user has been unfollowed.
        "400":
          description: |-
            The unfollow request is ill-formed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: |-
            The user identified by this name does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: |-
            The user is not being followed by the user specified in the body of the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: |-
            The user is not correctly authenticated (the given ID does not match the user's ID).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /resources/photos/{UUID}:
    parameters:
      - name: UUID
        in: path
        description: The photo's UUID
        required: true
        schema:
          $ref: "#/components/schemas/SHA256hash"
    get:
      operationId: getPhoto
      summary: Get a photo
      description: |-
        Get a photo
      tags:
        - "photos"
      responses:
        "200":
          description: |-
            The photo
          content:
            image/jpeg:
              schema:
                $ref: "#/components/schemas/Photo"
        "404":
          description: |-
            The photo does not exist

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...